Adds better defaults to devenv up / process compose
Sometimes the TUI is buggy and it's fiddly to copy and paste logs/errors
from that interface. I've disabled json logging and added a `make logs`
wrapper that tails `.devenv/state/process-compose/process-compose.log`.
It's much more manageable now that the logs are plain text. It's easy to
copy paste and grep for stuff. Ie. you can `make logs | grep web` to
tail just the django logs.

There is also the possibility of disabling the TUI altogether which just
tails the logs automatically after you do `devenv up`.
I've not yet decided if I want this to be the default.
The TUI (with all it's faults) is still useful for filtering logs
(ie. Choose the processes you want to look at) and restarting (Ctrl + R)
certain processes without needing to restart everything.